Fix calculating the total audio duration
At least, for the first encountered podcast episode. It has the duration
of 4263.6 s whereas the parses used to return 4262.1 s.

+++ src/MP3.hs
@@ -19,7 +19,7 @@ import ID3V1 qualified
 import Text.Printf
 
 -- | Duration of an MP3 file, in seconds.
-newtype AudioDuration = AudioDuration { getAudioDuration :: Float }
+newtype AudioDuration = AudioDuration { getAudioDuration :: Double }
   deriving newtype (Eq, Ord, Fractional, Num)
 
 instance Show AudioDuration where
